Added more entities

+++ src/main/java/com/example/demo/repository/BaseRepository.java
@@ -3,6 +3,6 @@ package com.example.demo.repository;
 import com.example.demo.model.Base;
 import org.springframework.data.repository.CrudRepository;
 
-public interface BaseRepository extends CrudRepository<Base, Long> {
+public interface BaseRepository extends CrudRepository<Base, Integer> {
 
 }
